requirements.txt
yfinance
ta
python-telegram-bot==13.15
